Automate Secret Santa Emails with SMTP
Secret Santa is a not-to-be-missed holiday tradition amongst friends, families, and coworkers where everyone draws random names to become someone else's "Secret Santa". The Secret Santa is then given a wish list of gift ideas to give to their chosen giftee.

After opening their present, the giftee has to guess which person was their Secret Santa. It’s a holiday classic where the guessing is as much a part of the fun as receiving the gift.